Split up + move entity plugins
